change name "notebook-ssh-controller" to "notebook-external-access-controller", add external-access controller
