+++ cgu/notebook-external-access-controller/config/manager/kustomization.yaml
@@ -1,0 +1,8 @@
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+resources:
+- manager.yaml
+
+images:
+- name: controller
+  newName: cguaicadmin/notebook-external-access-controller:V1.1.1
