+++ cgu/notebook-external-access-controller/config/manager/manager.yaml
@@ -1,0 +1,68 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: controller-manager
+  namespace: cgu
+  labels:
+    control-plane: notebook-external-access-controller
+spec:
+  selector:
+    matchLabels:
+      control-plane: notebook-external-access-controller
+  replicas: 1
+  template:
+    metadata:
+      labels:
+        control-plane: notebook-external-access-controller
+    spec:
+      securityContext:
+        runAsNonRoot: true
+      containers:
+      - command:
+        - /manager
+        args:
+        - --metrics-bind-address=127.0.0.1:8080
+        - --health-probe-bind-address=:8081
+        - --leader-elect
+        image: cguaicadmin/notebook-external-access-controller:V1.1.1
+        imagePullPolicy: Always
+        name: manager
+        ports:
+        - containerPort: 9443
+          name: webhook-server
+          protocol: TCP
+        volumeMounts:
+        - mountPath: /tmp/k8s-webhook-server/serving-certs
+          name: cert
+          readOnly: true
+        securityContext:
+          allowPrivilegeEscalation: false
+          capabilities:
+            drop:
+              - ALL
+        livenessProbe:
+          httpGet:
+            path: /healthz
+            port: 8081
+          initialDelaySeconds: 15
+          periodSeconds: 20
+        readinessProbe:
+          httpGet:
+            path: /readyz
+            port: 8081
+          initialDelaySeconds: 5
+          periodSeconds: 10
+        resources:
+          limits:
+            cpu: 100m
+            memory: 100Mi
+          requests:
+            cpu: 10m
+            memory: 64Mi
+      serviceAccountName: default # 使用 'default' ServiceAccount
+      terminationGracePeriodSeconds: 10
+      volumes:
+      - name: cert
+        secret:
+          defaultMode: 420
+          secretName: webhook-server-cert
