+++ cgu/notebook-external-access-controller/config/rbac/role_binding.yaml
@@ -1,0 +1,12 @@
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: notebook-external-access-controller-manager-rolebinding
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: notebook-external-access-controller-manager-role
+subjects:
+- kind: ServiceAccount
+  name: default
+  namespace: cgu # 假設 Controller 部署在 'system' 命名空間
