--- cgu/notebook-ssh-controller/Dockerfile
@@ -1,25 +1,0 @@
-# Build the manager binary
-FROM golang:1.23 as builder
-
-WORKDIR /workspace
-# Copy the Go Modules manifests
-COPY go.mod go.mod
-COPY go.sum go.sum
-# cache deps before building and copying source code
-# they don't change often, so they are cached in a layer
-RUN go mod download
-
-# Copy the go source
-COPY main.go main.go
-COPY controllers/ controllers/
-
-# Build
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -a -o manager main.go
-
-# Use distroless as base image to reduce container size
-FROM gcr.io/distroless/static:nonroot
-WORKDIR /
-COPY --from=builder /workspace/manager .
-USER 65532:65532
-
-ENTRYPOINT ["/manager"]
