--- cgu/notebook-ssh-controller/Makefile
@@ -1,32 +1,0 @@
-# Image URL to use all building/pushing image targets
-IMG ?= notebook-ssh-controller
-VER ?= 1.1.0
-# Build the docker image
-.PHONY: docker-build
-docker-build:
-	docker build -t cguaicadmin/${IMG}:${VER} .
-
-# Push the docker image
-.PHONY: docker-push
-docker-push:
-	docker push cguaicadmin/${IMG}:${VER}
-
-# Deploy to the cluster
-.PHONY: deploy
-deploy:
-	kubectl apply -k config/default
-
-# Undeploy from the cluster
-.PHONY: undeploy
-undeploy:
-	kubectl delete -k config/default
-
-# Run tests
-.PHONY: test
-test:
-	go test ./... -coverprofile cover.out
-
-# Clean up
-.PHONY: clean
-clean:
-	rm -f manager cover.out
