--- cgu/notebook-ssh-controller/OperatorPattern.md
@@ -1,69 +1,0 @@
-# Kubernetes Operator 中的控制器設定解析
-
-這份文件旨在解釋位於 `controllers/notebook_controller.go:149-155` 的 Go 程式碼。這段程式碼是使用 [Kubebuilder](https://book.kubebuilder.io/) 框架（其底層為 [controller-runtime](https://github.com/kubernetes-sigs/controller-runtime)）來設定一個 Kubernetes 控制器。
-
-```go
-func (r *NotebookReconciler) SetupWithManager(mgr ctrl.Manager) error {
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&notebookv1.Notebook{}).
-		Owns(&corev1.Service{}).
-		Complete(r)
-}
-```
-
----
-
-### 1. 目的與功能 (Purpose and Functionality)
-
-這段程式碼的 **主要目的** 是將 [`NotebookReconciler`](controllers/notebook_controller.go:149)（控制器邏輯的實作）註冊到 `Manager` 中。`Manager` 是運行一個或多個控制器的核心程序。
-
-簡而言之，這段程式碼在做以下設定：
-
-*   建立一個新的控制器。
-*   告訴控制器要「監看」[`Notebook`](controllers/notebook_controller.go:152) 這種自訂資源（Custom Resource）的變化。
-*   同時，也告訴控制器它「擁有」並管理 [`Service`](controllers/notebook_controller.go:153) 這種原生的 Kubernetes 資源。
-*   最後，將這個設定好的控制器與 [`NotebookReconciler`](controllers/notebook_controller.go:149) 的協調（reconcile）邏輯綁定在一起。
-
-一旦設定完成，每當有 `Notebook` 資源被建立、更新或刪除，或是由 `Notebook` 所擁有的 `Service` 資源發生變化時，控制器的 `Reconcile` 函式就會被觸發，以確保系統的實際狀態符合使用者在 `Notebook` 資源中定義的期望狀態。
-
----
-
-### 2. 關鍵元件與其互動 (Key Components and Their Interactions)
-
-讓我們逐行拆解這段程式碼：
-
-*   [`func (r *NotebookReconciler) SetupWithManager(mgr ctrl.Manager) error`](controllers/notebook_controller.go:149):
-    *   這是在 [`NotebookReconciler`](controllers/notebook_controller.go:149) 這個結構上定義的一個方法。`r` 代表 `Reconciler` 的實例，它包含了處理資源狀態協調的主要邏輯。
-    *   `mgr ctrl.Manager`: 這是控制器管理器，是整個 Operator 的核心。它負責與 Kubernetes API Server 溝通、管理快取（Cache）、執行控制器等。
-
-*   [`return ctrl.NewControllerManagedBy(mgr).`](controllers/notebook_controller.go:150):
-    *   這會建立一個新的控制器 **建構器 (Builder)**。所有後續的設定都會在這個建構器的基礎上進行鏈式呼叫。控制器將由傳入的 `mgr` 來管理。
-
-*   [`.For(&notebookv1.Notebook{}).`](controllers/notebook_controller.go:152):
-    *   這是最重要的設定之一。它指定了控制器監看的 **主要資源**。
-    *   這行程式碼告訴控制器：「請監聽所有 `notebookv1.Notebook` 類型的資源。任何關於它們的事件（建立、更新、刪除）都應該觸發一次協調（Reconciliation）。」
-
-*   [`.Owns(&corev1.Service{}).`](controllers/notebook_controller.go:153):
-    *   這指定了控制器 **擁有 (Owns)** 的 **次要資源**。
-    *   在 Operator 模式中，一個控制器通常會建立和管理其他資源。例如，`Notebook` 控制器可能會為每個 `Notebook` 實例建立一個對應的 `Service`。
-    *   這行設定確保了：當一個**由 `Notebook` 擁有的 `Service`** 發生變化時，也會觸發其 **擁有者 `Notebook`** 的協調。這是一種強大的機制，可以讓控制器在它所管理的子資源被意外修改或刪除時，能自動進行修復。
-
-*   [`.Complete(r)`](controllers/notebook_controller.go:154):
-    *   這是建構過程的最後一步，它完成了控制器的建立。
-    *   參數 `r` 就是 `NotebookReconciler` 的實例。這一步等於是將前面定義的所有事件觸發規則與 `r` 內部的 `Reconcile` 方法（實際的業務邏輯）連結起來。
-
----
-
-### 3. 重要模式或技術 (Important Patterns or Techniques)
-
-*   **Operator 模式 (Operator Pattern)**:
-    這段程式碼是實現 Kubernetes Operator 的典範。Operator 是一種特殊的控制器，它利用自訂資源（CRD）來管理複雜的應用程式及其生命週期。
-
-*   **協調循環 (Reconciliation Loop)**:
-    這是所有 Kubernetes 控制器的核心概念。控制器不斷地比較資源的「期望狀態」（定義在 `Notebook` 物件中）和「實際狀態」（Kubernetes 叢集中的真實情況），並採取行動來彌補兩者之間的差異。這段程式碼定義了觸發這個循環的條件。
-
-*   **擁有者參考 (Owner References)**:
-    `.Owns()` 的功能是基於 Kubernetes 的 `OwnerReference` 機制。當 `Notebook` 控制器建立一個 `Service` 時，它會在 `Service` 的元數據（metadata）中設定一個指向該 `Notebook` 的 `ownerReference`。這不僅能實現級聯刪除（當 `Notebook` 被刪除時，其擁有的 `Service` 也會被自動清理），還能像這裡一樣，讓子資源的變化能夠反向通知父資源的控制器。
-
-*   **建構器模式 (Builder Pattern)**:
-    這種 `New...().For().Owns().Complete()` 的鏈式呼叫寫法是建構器模式的體現。它讓建立一個複雜物件（此處為控制器）的過程變得聲明式 (declarative) 且易於閱讀和理解。
