--- cgu/notebook-ssh-controller/config/certmanager/certificate.yaml
@@ -1,15 +1,0 @@
-# The following manifests contain a self-signed issuer CR and a certificate CR.
-# More document can be found at https://docs.cert-manager.io
-apiVersion: cert-manager.io/v1
-kind: Certificate
-metadata:
-  name: notebook-serving-cert
-  namespace: cgu
-spec:
-  dnsNames:
-  - notebook-ssh-notebook-webhook-service.cgu.svc
-  - notebook-ssh-notebook-webhook-service.cgu.svc.cluster.local
-  issuerRef:
-    kind: ClusterIssuer
-    name: kubeflow-self-signing-issuer
-  secretName: webhook-server-cert
