--- cgu/notebook-ssh-controller/config/certmanager/kustomization.yaml
@@ -1,4 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-resources:
-- certificate.yaml
