--- cgu/notebook-ssh-controller/config/default/kustomization.yaml
@@ -1,17 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-
-# Adds namespace to all resources.
-namespace: cgu
-
-# Value of this field is prepended to the
-# names of all resources, e.g. a deployment named
-# "wordpress" becomes "alices-wordpress".
-# Note that it should also match with the prefix (text before '-') of the namespace
-# field above.
-namePrefix: notebook-ssh-
-
-resources:
-- ../manager
-- ../webhook
-- ../certmanager
