--- cgu/notebook-ssh-controller/config/manager/kustomization.yaml
@@ -1,8 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-resources:
-- manager.yaml
-
-images:
-- name: controller
-  newName: cguaicadmin/notebook-ssh-controller:V1.1.1
