--- cgu/notebook-ssh-controller/config/rbac/role_binding.yaml
@@ -1,12 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: notebook-ssh-controller-manager-rolebinding
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: notebook-ssh-controller-manager-role
-subjects:
-- kind: ServiceAccount
-  name: default
-  namespace: cgu # 假設 Controller 部署在 'system' 命名空間
