--- cgu/notebook-ssh-controller/config/webhook/kustomization.yaml
@@ -1,5 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-resources:
-- manifests.yaml
-- service.yaml
