--- cgu/notebook-ssh-controller/config/webhook/manifests.yaml
@@ -1,27 +1,0 @@
-apiVersion: admissionregistration.k8s.io/v1
-kind: ValidatingWebhookConfiguration
-metadata:
-  name: notebook-validating-webhook-configuration
-  annotations:
-    cert-manager.io/inject-ca-from: cgu/notebook-ssh-notebook-serving-cert
-webhooks:
-- admissionReviewVersions:
-  - v1
-  clientConfig:
-    service:
-      name: notebook-ssh-notebook-webhook-service
-      namespace: cgu
-      path: /validate-kubeflow-org-v1beta1-notebook
-  failurePolicy: Fail
-  name: vnotebook.kb.io
-  rules:
-  - apiGroups:
-    - kubeflow.org
-    apiVersions:
-    - v1beta1
-    operations:
-    - CREATE
-    - UPDATE
-    resources:
-    - notebooks
-  sideEffects: None
