--- cgu/notebook-ssh-controller/config/webhook/service.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: notebook-webhook-service
-  namespace: cgu
-spec:
-  ports:
-  - port: 443
-    protocol: TCP
-    targetPort: 9443
-  selector:
-    control-plane: controller-manager
